desiholmes/requirements.txt
numpy>=1.22.0
pandas>=1.4.0
scikit-learn>=1.0.2
tensorflow>=2.8.0
torch>=1.11.0
transformers>=4.18.0
spacy>=3.2.0
networkx>=2.7.1
matplotlib>=3.5.1
plotly>=5.6.0
fastapi>=0.75.0
pydantic>=1.9.0
python-multipart>=0.0.5
uvicorn>=0.17.6
pillow>=9.0.1